# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/timesheets:
    get:
      tags:
        - Timesheets
      summary: List Timesheets
      description: Get a paginated list of Timesheets with optional filtering and sorting
      operationId: listTimesheets
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimesheetList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Timesheets
      summary: Create Timesheet
      description: Create a new Timesheet
      operationId: createTimesheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimesheetRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Timesheet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/timesheet_approvals:
    get:
      tags:
        - TimesheetApprovals
      summary: List TimesheetApprovals
      description: Get a paginated list of TimesheetApprovals with optional filtering and sorting
      operationId: listTimesheetApprovals
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimesheetApprovalList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - TimesheetApprovals
      summary: Create TimesheetApproval
      description: Create a new TimesheetApproval
      operationId: createTimesheetApproval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimesheetApprovalRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimesheetApproval'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Timesheet:
      type: object
      required:
        - id
        - company_id
        - employee_id
        - year
        - month
        - date
        - entry_type
        - unit_amount
        - currency
        - is_billable
        - billable_amount
        - costing_amount
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        project_id:
          nullable: true
          type: string
          format: uuid
        task_id:
          nullable: true
          type: string
          format: uuid
        year:
          type: integer
          format: int32
        month:
          type: integer
          format: int32
        date:
          type: string
          format: date
        remark:
          nullable: true
          type: string
        time_start:
          nullable: true
          type: string
          format: date-time
        time_end:
          nullable: true
          type: string
          format: date-time
        entry_type:
          type: string
        unit_amount:
          type: string
        currency:
          type: string
        activity_type_id:
          nullable: true
          type: string
          format: uuid
        billing_rate:
          nullable: true
          type: string
        costing_rate:
          nullable: true
          type: string
        is_billable:
          type: boolean
        billable_amount:
          type: string
        costing_amount:
          type: string
        invoice_id:
          nullable: true
          type: string
          format: uuid
        source_timeoff_request_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateTimesheetRequest:
      type: object
      required:
        - company_id
        - employee_id
        - year
        - month
        - date
        - entry_type
        - unit_amount
        - currency
        - is_billable
        - billable_amount
        - costing_amount
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        project_id:
          nullable: true
          type: string
          format: uuid
        task_id:
          nullable: true
          type: string
          format: uuid
        year:
          type: integer
          format: int32
        month:
          type: integer
          format: int32
        date:
          type: string
          format: date
        remark:
          nullable: true
          type: string
        time_start:
          nullable: true
          type: string
          format: date-time
        time_end:
          nullable: true
          type: string
          format: date-time
        entry_type:
          type: string
        unit_amount:
          type: string
        currency:
          type: string
        activity_type_id:
          nullable: true
          type: string
          format: uuid
        billing_rate:
          nullable: true
          type: string
        costing_rate:
          nullable: true
          type: string
        is_billable:
          type: boolean
        billable_amount:
          type: string
        costing_amount:
          type: string
        invoice_id:
          nullable: true
          type: string
          format: uuid
        source_timeoff_request_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    TimesheetList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Timesheet'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    TimesheetApproval:
      type: object
      required:
        - id
        - company_id
        - employee_id
        - year
        - month
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        approver_id:
          nullable: true
          type: string
          format: uuid
        year:
          type: integer
          format: int32
        month:
          type: integer
          format: int32
        remark:
          nullable: true
          type: string
        billable_time:
          nullable: true
          type: string
        billable_cost:
          nullable: true
          type: string
        status:
          type: string
        approval_request_id:
          nullable: true
          type: string
          format: uuid
        submitted_at:
          nullable: true
          type: string
          format: date-time
        data:
          nullable: true
          type: object
        metadata:
          type: object
...
    CreateTimesheetApprovalRequest:
      type: object
      required:
        - company_id
        - employee_id
        - year
        - month
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        approver_id:
          nullable: true
          type: string
          format: uuid
        year:
          type: integer
          format: int32
        month:
          type: integer
          format: int32
        remark:
          nullable: true
          type: string
        billable_time:
          nullable: true
          type: string
        billable_cost:
          nullable: true
          type: string
        status:
          type: string
        approval_request_id:
          nullable: true
          type: string
          format: uuid
        submitted_at:
          nullable: true
          type: string
          format: date-time
        data:
          nullable: true
          type: object
        metadata:
          type: object
...
    TimesheetApprovalList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TimesheetApproval'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []