# src/exports/types.rs
//! Public types for Timesheet module
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! These DTOs are the ONLY types other modules should use.
//! They are decoupled from internal domain entities.

use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{DateTime, Utc, NaiveDate};
use rust_decimal::Decimal;
use crate::domain::entity::*;

// ============================================================================
// TIMESHEET TYPES
// ============================================================================

/// Type-safe ID for Timesheet
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct TimesheetId(pub Uuid);

impl TimesheetId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for TimesheetId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<TimesheetId> for Uuid {
    fn from(id: TimesheetId) -> Self {
        id.0
    }
}

/// Data transfer object for Timesheet
///
/// This is the public representation of Timesheet for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct TimesheetDto {
    pub id: TimesheetId,
    pub company_id: Uuid,
    pub employee_id: Uuid,
    pub project_id: Option<Uuid>,
    pub task_id: Option<Uuid>,
    pub year: i32,
    pub month: i32,
    pub date: NaiveDate,
    pub remark: Option<String>,
    pub time_start: Option<DateTime<Utc>>,
    pub time_end: Option<DateTime<Utc>>,
    pub entry_type: TimesheetType,
    pub unit_amount: Decimal,
    pub currency: String,
    pub activity_type_id: Option<Uuid>,
    pub billing_rate: Option<Decimal>,
    pub costing_rate: Option<Decimal>,
    pub is_billable: bool,
    pub billable_amount: Decimal,
    pub costing_amount: Decimal,
    pub invoice_id: Option<Uuid>,
    pub source_timeoff_request_id: Option<Uuid>,
    pub metadata: serde_json::Value,
}

/// Summary view of Timesheet for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct TimesheetSummary {
    pub id: TimesheetId,
}

/// Reference to Timesheet for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct TimesheetRef {
    pub id: TimesheetId,
}

// ============================================================================
// TIMESHEETAPPROVAL TYPES
// ============================================================================

/// Type-safe ID for TimesheetApproval
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct TimesheetApprovalId(pub Uuid);

impl TimesheetApprovalId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for TimesheetApprovalId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<TimesheetApprovalId> for Uuid {
    fn from(id: TimesheetApprovalId) -> Self {
        id.0
    }
}

/// Data transfer object for TimesheetApproval
///
/// This is the public representation of TimesheetApproval for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct TimesheetApprovalDto {
    pub id: TimesheetApprovalId,
    pub company_id: Uuid,
    pub employee_id: Uuid,
    pub approver_id: Option<Uuid>,
    pub year: i32,
    pub month: i32,
    pub remark: Option<String>,
    pub billable_time: Option<Decimal>,
    pub billable_cost: Option<Decimal>,
    pub status: TimesheetApprovalStatus,
    pub approval_request_id: Option<Uuid>,
    pub submitted_at: Option<DateTime<Utc>>,
    pub data: Option<serde_json::Value>,
    pub metadata: serde_json::Value,
}

/// Summary view of TimesheetApproval for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct TimesheetApprovalSummary {
    pub id: TimesheetApprovalId,
    pub status: TimesheetApprovalStatus,
}

/// Reference to TimesheetApproval for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct TimesheetApprovalRef {
    pub id: TimesheetApprovalId,
}

// ============================================================================
// CUSTOM TYPES
// ============================================================================

// <<< CUSTOM TYPES START >>>
// Add custom public types here
// <<< CUSTOM TYPES END >>>